requirements.txt
django
django-allauth
django-bootstrap-form
django-crispy-forms
django-environ
django-model-utils
django-phonenumber-field[phonenumbers]
django-positions
django-sekizai
django-reversion

-e git+https://github.com/iarp/Python-Utils.git#egg=iarp_utils
-e git+https://github.com/iarp/iarp-django-utils.git#egg=iarp_django_utils
-e git+https://github.com/iarp/django-templated-emailer.git#egg=django_templated_emailer

loguru
psycopg2-binary
requests
sentry-sdk
selenium
pytz
